Remove duplicate mask setup code
